Run multiple AirPlay servers.
The `srand()` call needed to come out because reseeding the random
number generator with the same seed caused both successive calls to
`random_mac()` to return the same randomized address. The call isn't
necessary in the first place: OBS does seed the RNG, but even if it
didn't, it's more important for random MACs generated within the same
process to be different from each other than it is for them to be unique
across subsequent program invocations.